+++ D3D12Demo/D3D12Demo/Common/MeshReader.h
@@ -9,5 +9,5 @@
 class MeshReader
 {
 public:
-	static void LoadFromTxt(const std::string filename, std::vector<Vertex> &vertices, std::vector<uint16_t> &indices);
+	static void LoadFromTxt(const std::string filename, std::vector<Vertex> &vertices, std::vector<uint32_t> &indices);
 };
